Deployement : proxy broker stream via nginx ; separate mqtt and http apis executions ; improve cluster mode process ( avoid repeated boot actions, fix scheduler regeneration )

+++ src/boot/11-init-cache.js
@@ -1,5 +1,9 @@
 module.exports = async function initCache(app) {
   try {
+    if (process.env.CLUSTER_MODE) {
+      if (process.env.PROCESS_ID !== '0') return null;
+      if (process.env.INSTANCES_PREFIX && process.env.INSTANCES_PREFIX !== '1') return null;
+    }
     const devices = await app.models.Device.syncCache('DOWN');
     const clients = await app.models.Client.deleteAll();
     // todo keep schedulers cache and find timers
